feat: Phase 3A - Component breakdown and shared utilities
- Extract JobCard, FilterPanel, and Pagination components from BrowseJobs
- Extract StatCard, TabButton, and ActivitySection components from AdminDashboard
- Create shared formatters utility with formatBudget, formatDuration, formatCategory
- Consolidate duplicate utility functions across components
- Maintain existing functionality and styling
- Improve code reusability and maintainability
- Reduce BrowseJobs from 313 to ~120 lines
- Extract inline components from AdminDashboard for better organization
